pyproject.toml
[project]
name = "pre-commit-hub"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "itsjohncs", email = "johnsullivan.pem@gmail.com" }
]
dependencies = [
    "pyyaml>=6.0.1",
    "requests>=2.32.3",
    "pydantic>=2.8.2",
    "thefuzz>=0.22.1",
    "pytest>=8.3.2",
    "pygithub>=2.3.0",
]
readme = "README.md"
requires-python = ">= 3.9"

[project.scripts]
"pre-commit-hub" = "pre_commit_hub:cli.main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pre_commit_hub"]

[tool.pyright]
venvPath = "."
venv = ".venv"
